% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_conjugue.R
\name{afficher_resultats}
\alias{afficher_resultats}
\title{Display Results of the Optimization}
\usage{
afficher_resultats(x)
}
\arguments{
\item{x}{Solution vector representing resource allocation}
}
\description{
This function displays the solution vector x representing the resource allocation
}
\examples{
# x <- c(5, 3, 4, 2, 6)
# afficher_resultats(x) # Affiche le vecteur solution x représentant les allocations de ressources

}
