% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_conjugue.R
\name{generer_matrice}
\alias{generer_matrice}
\title{Generate Coefficient Matrix for Queue Management}
\usage{
generer_matrice(n)
}
\arguments{
\item{n}{Number of resources (doctors)}
}
\value{
Coefficient matrix A
}
\description{
This function generates the coefficient matrix A for queue management based on the number of resources (doctors) n
}
\examples{
# n <- 5
# generer_matrice(n) # Génère une matrice A tridiagonale symétrique et définie positive

}
