% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_conjugue.R
\name{gradient_conjugue}
\alias{gradient_conjugue}
\title{Gradient Conjugate Method}
\usage{
gradient_conjugue(A, b, tol = 1e-08, max_iter = 1000)
}
\arguments{
\item{A}{Coefficient matrix}

\item{b}{Right-hand side vector}

\item{tol}{Tolerance for convergence}

\item{max_iter}{Maximum number of iterations}
}
\value{
Solution vector
}
\description{
This function implements the conjugate gradient method to solve the linear system Ax=b
}
\examples{
# A <- matrix(c(4, 1, 1, 3), nrow=2, byrow=TRUE)
# b <- c(1, 2)
# gradient_conjugue(A, b) # Résout le système Ax=b

}
