% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_conjugue.R
\name{tracer_graphique}
\alias{tracer_graphique}
\title{Plot Resource Allocation Graph}
\usage{
tracer_graphique(x)
}
\arguments{
\item{x}{Solution vector representing resource allocation}
}
\description{
This function draws a resource allocation graph
}
\examples{
# x <- c(5, 3, 4, 2, 6)
# tracer_graphique(x) # Trace un graphique du vecteur solution x

}
